Skip upload stats for test reports from rerun disabled tests workflow
